deploy create_visitors with env vars
