+++ create_visitors/deploy.sh
@@ -5,6 +5,7 @@
 # deploy cloud function
 gcloud functions deploy --region=$TARGET_REGION create_visitors \
 --runtime python38 \
+--env-vars-file=env_vars.yml \
 --trigger-resource $NEWGARDEN_VISITORS_CSV_DATA_BUCKET \
 --trigger-event google.storage.object.finalize
