+++ create_visitors/env_vars.yml
@@ -1,0 +1,5 @@
+SERVICE_ACCOUNT_CREDENTIALS_JSON: newgarden-cloud-functions-9f405fe88d29.json
+SERVICE_ACCOUNT_DATA_BUCKET: newgarden-service-account-data-bucket
+NEWGARDEN_VISITORS_SPREADSHEET_ID: 1NX2DQrM2WlVpZdUQHqoSy9jFQHiI-JlWtxCPaMVPKpk
+NEWGARDEN_VISITORS_HEADER_RANGE_NAME: Visitors!A1:AI1
+
